y get computed first

+++ The platformer/Player.h
@@ -18,6 +18,6 @@ private:
 	bool m_jumping = false;
 	Animation m_animation;
 	RigidBody m_body;
-	sf::Vector2f m_lastSafePosition;
+	sf::Vector2i m_lastSafePosition;
 };
